Add 'only' parameter to Action
